requirements.txt
flask=3.0.3
flask_restful=0.3.10
sqlalchemy=2.0.36
pymysql=1.1.1
cryptography=44.0.0
python-dotenv=1.0.1
PyJWT=2.9.0
Pillow=10.1.0
redis=5.2.1

